Move contracts to contracts.py
