+++ databuilder/backends/graphnet.py
@@ -1,4 +1,4 @@
-from ..contracts import tables
+from ..contracts import contracts
 from ..query_engines.mssql import MssqlQueryEngine
 from .base import BaseBackend, Column, MappedTable
 
@@ -11,7 +11,7 @@ class GraphnetBackend(BaseBackend):
     patient_join_column = "Patient_ID"
 
     patients = MappedTable(
-        implements=tables.PatientDemographics,
+        implements=contracts.PatientDemographics,
         source="TRE.Patients",
         columns=dict(
             sex=Column("varchar", source="Sex"),
@@ -21,7 +21,7 @@ class GraphnetBackend(BaseBackend):
     )
 
     clinical_events = MappedTable(
-        implements=tables.WIP_ClinicalEvents,
+        implements=contracts.WIP_ClinicalEvents,
         source="TRE.ClinicalEvents",
         columns=dict(
             code=Column("varchar", source="Code"),
@@ -32,7 +32,7 @@ class GraphnetBackend(BaseBackend):
     )
 
     practice_registrations = MappedTable(
-        implements=tables.WIP_PracticeRegistrations,
+        implements=contracts.WIP_PracticeRegistrations,
         source="TRE.PracticeRegistrations",
         columns=dict(
             pseudo_id=Column("integer", source="Organisation_ID"),
@@ -43,7 +43,7 @@ class GraphnetBackend(BaseBackend):
     )
 
     covid_test_results = MappedTable(
-        implements=tables.WIP_TestResults,
+        implements=contracts.WIP_TestResults,
         source="TRE.CovidTestResults",
         columns=dict(
             date=Column("date", source="SpecimenDate"),
@@ -52,7 +52,7 @@ class GraphnetBackend(BaseBackend):
     )
 
     hospitalizations = MappedTable(
-        implements=tables.WIP_HospitalizationsWithoutSystem,
+        implements=contracts.WIP_HospitalizationsWithoutSystem,
         source="TRE.Hospitalisations",
         columns=dict(
             date=Column("date", source="AdmitDate"),
@@ -61,7 +61,7 @@ class GraphnetBackend(BaseBackend):
     )
 
     patient_address = MappedTable(
-        implements=tables.WIP_PatientAddress,
+        implements=contracts.WIP_PatientAddress,
         source="TRE.PatientAddresses",
         columns=dict(
             patientaddress_id=Column("integer", source="PatientAddress_ID"),
